Provide an alternative configuration via ApplicationContextInitializer in tests.
This is a slightly different approach. this
ApplicationContextInitializer is configured via @ContextConfiguration(initializers = [RepositoryTest.Initializer::class])

It changes the configuration properties instead of defining a custom
bean.

While this is a bit more code, i like this approach better as it is as
close as possible to the production code.