Remove test harness
The test harness of course depends on Neo4j which has a lot of
dependencies. These take their toll on the classpath and in case of some
(Jersey and JAX-RS) are known to clash with the ones Spring Boot might
have.

If you don't have a good reason for having any of them on the classpath,
I would consider removing them.

We reocmmend test harness for testing stored procedures and the like.

If you can afford it runtime wise to bring up a container, I would go
that way.

A third option is to use Neo4j-OGM embedded with DataNeo4jTest but
that's essentially the OGM version of test harness and not "better" than
bring up the test harness alone.